@@ -13,12 +13,11 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2016": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
+    "babel-plugin-transform-export-extensions": "^6.22.0",
+    "babel-preset-env": "^1.6.0",
     "babel-runtime": "^6.23.0",
     "babelify": "^7.3.0",
+    "browserify-shim": "^3.8.14",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-browserify-multi-entry": "^1.0.2",
     "gulp-concat": "^2.6.0",
@@ -35,8 +34,22 @@
   },
   "dependencies": {
     "axios": "^0.16.2",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
-  "description": ""
+  "description": "",
+  "browserify": {
+    "transform": [
+        "babelify",
+        "reactify",
+        "browserify-shim"
+    ]
+  },
+  "browserify-shim": {
+    "shim": {
+        "react": "global:React",
+        "react-dom": "global:ReactDOM",
+        "axios": "global:axios"
+    }
+  }
 }
