@@ -1,8 +1,6 @@
 
-class Logo extends React.Component {
+export default class Logo extends React.Component {
     render() {
         return <img src="../assets/images/logo.png" alt="" className="logo" />;
     }
 }
-
-module.exports = Logo;
